+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Nightly](https://github.com/raymond-u/lungo/compare/v0.5.1...HEAD)
 
+### Fixed
+
+- Update user accounts correctly
+
 ## [0.5.1](https://github.com/raymond-u/lungo/compare/v0.5.0...v0.5.1) - 2024-6-22
 
 ### Fixed