+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Nightly](https://github.com/raymond-u/lungo/compare/v0.5.4...HEAD)
 
+### Fixed
+
+- Allow `plugins` in the configuration file to be omitted (try again)
+
 ## [0.5.4](https://github.com/raymond-u/lungo/compare/v0.5.3...v0.5.4) - 2024-6-26
 
 ### Fixed