+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Nightly](https://github.com/raymond-u/lungo/compare/v0.5.4...HEAD)
 
+### Changed
+
+- Update Xray to 1.8.16
+
 ### Fixed
 
 - Allow `plugins` in the configuration file to be omitted (try again)